Dynebolic 4.0 beta banner (#147)
* remove redundant class (see https://t.me/c/1682572031/14041)

* add d:b splash close to hero

commiting for reference; spent way too much time only to realize it
doesn't work on mobile. :'( wanted the svg is inline to follow the color
changeing on page-load, which comes at the cost of the text size on
small screens...

Probably the text needs to be separate from the image, but that means
formatting the whole thing in this magic non-CSS, CSS language that i
don't know.

* adds text to banner, removes background image

* feat: add software banner

---------

Co-authored-by: Giovanni Abbatepaolo <30571828+bbtgnn@users.noreply.github.com>